refactor: update u16 to u64
Refs: #1

+++ src/admin.rs
@@ -112,7 +112,7 @@ pub trait AdminModule: crate::config::ConfigModule + storage::StorageModule {
 
     // lock period (in days) and bond amount
     #[endpoint(setPeriodsBonds)]
-    fn set_lock_periods_and_bonds(&self, args: MultiValueEncoded<MultiValue2<u16, BigUint>>) {
+    fn set_lock_periods_and_bonds(&self, args: MultiValueEncoded<MultiValue2<u64, BigUint>>) {
         only_privileged!(self, ERR_NOT_PRIVILEGED);
         for input in args.into_iter() {
             let (lock_period, bond) = input.into_tuple();
